requirements.txt
django
awsebcli 
djangorestframework